% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungeData.R
\name{addPlayerID}
\alias{addPlayerID}
\title{prepAddPlayerID}
\usage{
addPlayerID(x_df, players_df)
}
\arguments{
\item{pos_proj}{data.frame of projection data}

\item{player_data}{data.frame of player information}

\item{teampos_data}{data.frame of team and position data}
}
\value{
data.frame containing player name, team, position
}
\description{
Add Lahman unique player IDs to MLB/FantasyPros data
}
