% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRelativeValues.R
\name{adjustObp}
\alias{adjustObp}
\title{#' @export}
\usage{
adjustObp(team_df)
}
\description{
simPitchSeason <- function(iter, 
                           rosterstats_df,
                           configs){
  
  # Build Teams
  set.seed(iter)
  rand_id <- rep(1:configs$nbr_owners, 
                 sum(configs$roster$count[configs$roster$type == 'pitch']))
  teams_ <- split(rosterstats_df, rand_id)
  
  teams_ <- lapply(teams_, adjustWhip)
  
  # Sum, flatten, adjust and return
  lapply(teams_, function(x) as.data.frame(t(colSums(x))))%>%
    dplyr::bind_rows() 
}
}
