% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customRankings.R
\name{adjustWhip}
\alias{adjustWhip}
\title{Adjust whip values by team numbers}
\usage{
adjustWhip(team_df)
}
\arguments{
\item{team_df}{Team of pitching stats}
}
\value{
Adjusted WHIP values
}
\description{
Adjust whip values by team numbers
}
