% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSeasons.R
\name{analyzeSimulations}
\alias{analyzeSimulations}
\title{Calculate SLG}
\usage{
analyzeSimulations(sim_objs)
}
\arguments{
\item{x_df}{Data.frame}
}
\value{
Vector of slugging percentages
}
\description{
Calculate slugging percentage
}
