% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{baseballref_df}
\alias{baseballref_df}
\title{Baseball Reference tables and codes}
\format{A \code{'tbl_df', 'tbl', "data.frame"} with 17 rows and 2 columns
 \describe{
  \item{data}{Type of data [chr]}
  \item{link_or_code}{html link or codes for functions [chr]}
}}
\source{
Author compiled
}
\usage{
data(baseballref_df)
}
\description{
Types of data from Baseball-reference.com (not complete)
}
\keyword{datasets}
