% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{batprojs_df}
\alias{batprojs_df}
\title{Batting Projections}
\format{A \code{"battingprojections", 'tbl_df', 'tbl', "data.frame"} data.frame with 
3,479 rows and 18 variables
\describe{
  \item{player}{Standard player name [fct]}
  \item{year}{Year of data [int]}
  \item{team}{Players team (ESPN code) [fct]}
  \item{pos}{Position(s) [char]}
  \item{ab}{At bats [int]}
  \item{h}{Hits [int]}
  \item{x2b}{Doubles [int]}
  \item{x3b}{Triples [int]}
  \item{r}{Runs [int]}
  \item{hr}{Home Runs [int]}
  \item{rbi}{Runs batted in [int]}
  \item{bb}{Base on balls (Walks) [int]}
  \item{sb}{Stolen bases [int]}
  \item{so}{Strikeouts [int]}
  \item{avg}{Batting average [dbl]}
  \item{obp}{On base percentage [dbl]}
  \item{slg}{Slugging percentage [dbl]}
  \item{ops}{On base percentage plus slugging [dbl]}
}}
\source{
Fantasy Pros Website
}
\usage{
data(batprojs_df)
}
\description{
Projected batting stats (Fantasy Pros)
}
\keyword{datasets}
