% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{batting_df}
\alias{batting_df}
\title{Basic batting statistics}
\format{A \code{"batting", 'tbl_df', 'tbl', "data.frame"} data.frame with 
6,211 rows and 21 variables
\describe{
  \item{player_id}{The unique Lahman player ID [chr]}
  \item{pa}{Plate Appearances [int]}
  \item{ab}{At bats [int]}
  \item{h}{hits [int]}
  \item{r}{Runs [int]}
  \item{hr}{Home Runs [int]}
  \item{rbi}{Runs batted in [int]}
  \item{bb}{Base on balls (Walks) [int]}
  \item{hbp}{Hit by pitch [int]}
  \item{ibb}{Intentional Walks [int]}
  \item{sb}{Stolen bases [int]}
  \item{so}{Strikeouts [int]}
  \item{x2b}{Doubles [int]}
  \item{x3b}{Triples [int]}
  \item{avg}{Batting average [num]}
  \item{obp}{On base percentage [num]}
  \item{slg}{Slugging percentage [num]}
  \item{gidp}{Ground into double plays [int]}
  \item{sf}{Sacrifice flies [int]}
  \item{sh}{Sacrifice hits (bunts) [int]}
  \item{year}{Year of data [int]}
}}
\source{
www.baseball-reference.com (via getBRStats(type = 'batting))
}
\usage{
data(batting_df)
}
\description{
Baseball reference batting stats: 2014 - 2018
}
\keyword{datasets}
