% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{batting_ndf}
\alias{batting_ndf}
\title{Basic batting statistics}
\format{A \code{"batting", 'tbl_df', 'tbl', "data.frame"} nested data.frame with 5,898
 rows and 3 (20) variables
\describe{
  \item{player_id}{The unique Lahman player ID [fct]}
  \item{year}{Year of data [int]}
  \item{batting_stats}{Nested data.frame of batting stats [list]}
  \item{...team}{Players team (Lahman code) [fct]}
  \item{...g}{Games at that position [int]}
  \item{...ab}{At bats [int]}
  \item{...r}{Runs [int]}
  \item{...hr}{Home Runs [int]}
  \item{...rbi}{Runs batted in [int]}
  \item{...sb}{Stolen bases [int]}
  \item{...so}{Strikeouts [int]}
  \item{...gidp}{Ground into double plays [int]}
  \item{...sh}{Sacrifice hits (bunts) [int]}
  \item{...sf}{Sacrifice flies [int]}
  \item{...h}{Hits [int]}
  \item{...x2b}{Doubles [int]}
  \item{...x3b}{Triples [int]}
  \item{...ibb}{Intentional Walks [int]}
  \item{...hbp}{Hit by pitch [int]}
  \item{...bb}{Base on balls (Walks) [int]}
  \item{...slg}{Slugging percentage [dbl]}
  \item{...pa}{Plate Appearances [int]}
  \item{...obp}{On base percentage [dbl]}
}}
\source{
Lahman baseball database
}
\usage{
data(batting_ndf)
}
\description{
Lahman Database batting stats: 2014 - 2017
}
\keyword{datasets}
