% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectData.R
\name{brExtractID}
\alias{brExtractID}
\title{Extract player ID}
\usage{
brExtractID(name_obj)
}
\arguments{
\item{name_obj}{Single line of stat code}
}
\value{
Player's BR ID
}
\description{
Get player ID from html code
}
