% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectData.R
\name{brExtractStats}
\alias{brExtractStats}
\title{Extract each individual stat}
\usage{
brExtractStats(stat_v)
}
\arguments{
\item{stat_v}{Single line of stat code}
}
\value{
One row data.frame with stat and value
}
\description{
Get a single stat name and value from HTML code
}
