% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectData.R
\name{brToTable}
\alias{brToTable}
\title{Convert Raw BR data to a table}
\usage{
brToTable(br_obj)
}
\arguments{
\item{br_obj}{Baseball reference scraped object (build internally in scrapeBRStats())}
}
\value{
Stats for each player in a key-value pair table
}
\description{
Convert scraped BR statistical information to a table
}
