% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customRankings.R
\name{calcRelativeRotisserieValue}
\alias{calcRelativeRotisserieValue}
\title{Calculate relative rotisserie value for a stat}
\usage{
calcRelativeRotisserieValue(data, name, rp, rv, configs)
}
\arguments{
\item{data}{Vector of stats or projections}

\item{name}{Name of stat}

\item{rp}{Replacement level value data.frame}

\item{rv}{Relative point value per stat}

\item{configs}{An `fbbConfigs` object}
}
\value{
Vector of relative values for a stat
}
\description{
Given the replacement level and the relative worth per category increase, 
 calculate the relative value of each of the stats/projections
}
