% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsCalculations.R
\name{calcSLG}
\alias{calcSLG}
\title{Calculate SLG}
\usage{
calcSLG(x_df)
}
\arguments{
\item{x_df}{Data.frame}
}
\value{
Vector of slugging percentages
}
\description{
Calculate slugging percentage
}
