% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSeasons.R
\name{categoryDiffEngine}
\alias{categoryDiffEngine}
\title{Analyze Season to get point differntials}
\usage{
categoryDiffEngine(x, metric = "median")
}
\arguments{
\item{season}{A `season` object}
}
\value{
Median Difference between rankings for category
}
\description{
Get point differentials across categories
}
