% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customRankings.R
\name{crExpertRVCorrections}
\alias{crExpertRVCorrections}
\title{Expert correction of rv values from simulations}
\usage{
crExpertRVCorrections(type, rv_df)
}
\arguments{
\item{type}{'hit' or 'pitch'}

\item{rv_df}{The raw category values from `seasonAnalysis`}
}
\value{
An corrected rv data.frame
}
\description{
As simulations results can be extreme (usually on the small end), this makes some
corrections to ensure the values are within reasonable ranges
}
