% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customRankings.R
\name{crRelativeBattingValue}
\alias{crRelativeBattingValue}
\title{Calculate relative rotisserie values for batters}
\usage{
crRelativeBattingValue(projs_df, batting_df, fielding_df, configs)
}
\arguments{
\item{projs_df}{Data.frame of batting projections}

\item{batting_df}{Data.frame of batting stats}

\item{fielding_df}{Data.frame of fielding stats}

\item{configs}{An `fbbConfigs` object}
}
\value{
An RRV object with custom batter values
}
\description{
Custom batter values based on relative rotisserie point values
}
