% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customRankings.R
\name{crRelativeRotisserieValue}
\alias{crRelativeRotisserieValue}
\title{Calculate relative rotisserie values}
\usage{
crRelativeRotisserieValue(projs_, stats_, configs, ...)
}
\arguments{
\item{projs_}{List of batting and pitching projections}

\item{stats_}{List of batting, fielding and pitching statistics}

\item{configs}{An `fbbConfigs` object}
}
\value{
An RRV object with custom player values
}
\description{
Custom player values based on relative rotisserie point values
}
