% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customRankings.R
\name{customRankings.fp}
\alias{customRankings.fp}
\title{Fantasy pros (fp)}
\usage{
\method{customRankings}{fp}(type, configs, verbose = 1)
}
\arguments{
\item{type}{Type of custom ranking system to use ('xxx' for now)}

\item{configs}{An `fbbConfigs` object}

\item{verbose}{[1] The level of reponse statements you'd like (higher = more)}
}
\value{
A custom ranking set
}
\description{
Custom rankings from fantasypros.com
}
