% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customRankings.R
\name{customRankings}
\alias{customRankings}
\title{Generic method for custom ranking processes}
\usage{
customRankings(type, configs, ...)
}
\arguments{
\item{type}{Type of custom ranking system to use ('rrv' for now)}

\item{configs}{An `fbbConfigs` object}

\item{...}{Additional arguments}

\item{verbose}{[1] The level of reponse statements you'd like (higher = more)}
}
\value{
A custom ranking set
}
\description{
Create custom rankings based on a variety of methods and means
}
