% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draftUpdateRankings}
\alias{draftUpdateRankings}
\title{Update a teams available rankings}
\usage{
draftUpdateRankings(rankings_df, picked_df, roster_df, configs)
}
\arguments{
\item{rankings_df}{The rankings data}

\item{picked_df}{Data on those players already picked}

\item{roster_df}{The teams current roster}

\item{configs}{An `fbbConfigs` object}
}
\value{
A custom ranking set
}
\description{
Update a teams available rankings
}
