% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{fielding_df}
\alias{fielding_df}
\title{Basic fielding statistics}
\format{A \code{"fielding", 'tbl_df', 'tbl', "data.frame"} data.frame with 
6,729 rows and 7 variables
\describe{
  \item{player_id}{The unique Lahman player ID [chr]}
  \item{g}{Games played [int]}
  \item{gs}{Games started [int]}
  \item{cg}{Complete games [int]}
  \item{innings}{Games at that position [num]}
  \item{e}{Errors at that position [int]}
  \item{year}{Year of stats [int]}
}}
\source{
www.baseball-reference.com (via getBRStats(type = 'fielding))
}
\usage{
data(fielding_df)
}
\description{
Baseball reference fielding stats: 2014 - 2018
}
\keyword{datasets}
