% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{fielding_ndf}
\alias{fielding_ndf}
\title{Basic fielding statistics}
\format{A \code{"fielding", 'tbl_df', 'tbl', "data.frame"} nested data.frame with 7,749 rows and 3 (4) variables
\describe{
  \item{player_id}{The unique Lahman player ID [fct]}
  \item{year}{Year of data [int]}
  \item{fielding_stats}{Nested data.frame of fielding stats [list]}
  \item{...team}{Players team (Lahman code) [fct]}
  \item{...pos}{Position played [fct]}
  \item{...g}{Games at that position [int]}
  \item{...e}{Errors at that position [int]}
}}
\source{
Lahman baseball database
}
\usage{
data(fielding_ndf)
}
\description{
Lahman Database fielding stats: 2014 - 2017
}
\keyword{datasets}
