% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customRankings.R
\name{findReplacementValue}
\alias{findReplacementValue}
\title{Find replacement level value for a given category}
\usage{
findReplacementValue(s_df, stats_v, rep_nbr)
}
\arguments{
\item{s_df}{stat information data.frame}

\item{stats_v}{A vector of the stat category values}

\item{rep_nbr}{THe number of the replacement players}
}
\value{
A scalar value representing the replacement value of that stat
}
\description{
Find the nth value of a category where that Nth level is considered the replacement 
player value
}
