% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMunging.R
\name{fixFPName}
\alias{fixFPName}
\title{Fix Fantasy Pros Name Field}
\usage{
fixFPName(df, field = "name", type = "Position")
}
\arguments{
\item{field}{['name'] Name of existing field to convert}

\item{type}{[Position] Includes the position as well as name + team}

\item{x}{Vector of existing names in FP format}
}
\value{
Data.frame of converted names, team and position
}
\description{
Extract out the player, pos and team from Fantasy Pros team names
}
