% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseballreferenceScraper.R
\name{getBRStats.pitching}
\alias{getBRStats.pitching}
\title{Wrapper for getting basic BR pitching stats}
\usage{
\method{getBRStats}{fielding}(type, year)
}
\arguments{
\item{type}{('fielding', 'pitching' or 'batting')}

\item{year}{year of stats}
}
\value{
Stats for that year
}
\description{
Get all necessary pitching stats
}
