% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseballreferenceScraper.R
\name{getBRStats}
\alias{getBRStats}
\title{Wrapper for getting BR stats (Generic Method)}
\usage{
getBRStats(type, year)
}
\arguments{
\item{type}{('fielding', 'pitching' or 'batting')}

\item{year}{year of stats}
}
\value{
Stats for that year
}
\description{
Get all necessary stats of one of three types (fielding, pitching, batting)
}
