% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{marcel2018_proj}
\alias{marcel2018_proj}
\title{Marcel projections from Baseball Reference}
\format{A list class \code{"marcelProjections"} containing two data.frames.
\describe{
  \item{batting}{ A \code{'battingProj', 'tbl_df', 'tbl', "data.frame"}  with 506 
  rows and 17 columns}
  \item{...ab}{At bats [int]}
  \item{...h}{Hits [int]}
  \item{...r}{Runs [int]}
  \item{...hr}{Home runs [int]}
  \item{...rbi}{Runs batted in [int]}
  \item{...bb}{Walks [int]}
  \item{...avg}{Batting average [num]}
  \item{...obp}{On base percentage [num]}
  \item{...slg}{Slugging Percentage [num]}
  \item{...so}{Strikeouts [int]}
  \item{...hbp}{hit by pitch [int]}
  \item{...ibb}{Intentional walks [int]}
  \item{...sb}{Stolen bases [int]}
  \item{...sf}{Sacrifice flies [int]}
  \item{...sh}{Sacrifice hits [int]}
  \item{...gidp}{Grounded into double plays [int]}
  \item{...reliability}{Percent of forecast due to player [num]}
  \item{pitching}{A \code{'pitchingProj', 'tbl_df', 'tbl', 'data.frame'} with 611 
  rows and 9 columns.}
  \item{...ip}{Innings Pitched [int]}
  \item{...k}{Strikeouts [int]}
  \item{...bb}{Walks [int]}
  \item{...era}{Earned Run Average [num]}
  \item{...h}{Hits [int]}
  \item{...r}{Runs [int]}
  \item{...sv}{Saves [int]}
  \item{...whip}{Walks + Hits / IP [num]}
  \item{...reliability}{Percent of forecast due to player [num]}
}}
\source{
Baseball Reference
}
\usage{
data(marcel2018_proj)
}
\description{
2018 Marcel Projections from Baseball-Reference.com.  Batting and pitching
}
\keyword{datasets}
