% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSeasons.R
\name{multiFBBSimulations}
\alias{multiFBBSimulations}
\title{Calculate SLG}
\usage{
multiFBBSimulations(sims, nbr_owners = 8, season_year = 2018,
  rankings = NULL, strategies = NULL, verbose = 1)
}
\arguments{
\item{x_df}{Data.frame}
}
\value{
Vector of slugging percentages
}
\description{
Calculate slugging percentage
}
