% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCollection.R
\name{organizeRSAnnBatting}
\alias{organizeRSAnnBatting}
\title{Organize Batting Data}
\usage{
organizeRSAnnBatting(x)
}
\arguments{
\item{x}{Raw retrosheet htm table row}
}
\value{
Data.frame of annual batting stats
}
\description{
Organize retrosheet raw batting data annual
}
