% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCollection.R
\name{organizeRSAnnFielding}
\alias{organizeRSAnnFielding}
\title{Organize Fielding Data}
\usage{
organizeRSAnnFielding(x)
}
\arguments{
\item{x}{Raw retrosheet htm table row}
}
\value{
Data.frame of annual fielding stats
}
\description{
Organize retrosheet raw fielding data annual
}
