% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrosheetScraper.R
\name{organizeRSAnnPitching}
\alias{organizeRSAnnPitching}
\title{Organize Pitching Data}
\usage{
organizeRSAnnPitching(x)
}
\arguments{
\item{x}{Raw retrosheet htm table row}
}
\value{
Data.frame of annual piching stats
}
\description{
Organize retrosheet raw pitching data annual
}
