% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{pitching_df}
\alias{pitching_df}
\title{Basic pitching statistics}
\format{A \code{"pitching", 'tbl_df', 'tbl', "data.frame"} data.frame with 
3,723 rows and 11 variables
\describe{
  \item{player_id}{The unique Lahman player ID [chr]}
  \item{ip}{Innings pitched [num]}
  \item{g}{Games [int]}
  \item{gs}{Games started [int]}
  \item{k}{Strikeouts [int]}
  \item{sv}{Saves [int]}
  \item{w}{Wins [int]}
  \item{whip}{(Walks + Hits) / Innings Pitched [num]}
  \item{qs}{Quality starts [int]}
  \item{holds}{Holds [int]}
  \item{year}{Year of data [int]}
}}
\source{
www.baseball-reference.com (via getBRStats(type = 'pitching))
}
\usage{
data(pitching_df)
}
\description{
Baseball reference pitching stats: 2014 - 2018
}
\keyword{datasets}
