% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{pitching_ndf}
\alias{pitching_ndf}
\title{Basic pitching statistics}
\format{A \code{"pitching", 'tbl_df', 'tbl', "data.frame"} nested data.frame with 2,924 rows and 3 (13) variables
\describe{
  \item{player_id}{The unique Lahman player ID [fct]}
  \item{year}{Year of data [int]}
  \item{pitching_stats}{Nested data.frame of pitching stats [list]}
  \item{...team}{Players team (Lahman code) [fct]}
  \item{...w}{Wins [int]}
  \item{...g}{Games [int]}
  \item{...gs}{Games started [int]}
  \item{...h}{Hits allowed [int]}
  \item{...er}{Earned runs allowed [int]}
  \item{...bb}{Base on balls issued [int]}
  \item{...k}{Strikeouts [int]}
  \item{...ibb}{Intentional Walks [int]}
  \item{...sv}{Saves [int]}
  \item{...ip}{Innings pitched [int]}
  \item{...whip}{(Walks + Hits) / Innings Pitched [dbl]}
}}
\source{
Lahman baseball database
}
\usage{
data(pitching_ndf)
}
\description{
Lahman Database pitching stats: 2014 - 2017
}
\keyword{datasets}
