% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{pitchprojs_df}
\alias{pitchprojs_df}
\title{Pitching Projections}
\format{A \code{"pitchingprojections", 'tbl_df', 'tbl', "data.frame"} data.frame with 
4,364 rows and 17 variables
\describe{
  \item{player}{Standard player name [fct]}
  \item{year}{Year of data [int]}
  \item{team}{Players team (ESPN code) [fct]}
  \item{g}{Games [int]}
  \item{gs}{Games started [int]}
  \item{cg}{Complete games [int]}
  \item{sv}{Saves [int]}
  \item{ip}{Innings pitched [int]}
  \item{w}{Wins [int]}
  \item{l}{Losses [int]}
  \item{h}{Hits allowed [int]}
  \item{er}{Earned runs allowed [int]}
  \item{k}{Strikeouts [int]}
  \item{bb}{Base on balls issued [int]}
  \item{era}{Earned Run Average [dbl]}
  \item{whip}{(Walks + Hits) / Innings Pitched [dbl]} 
  \item{player_id}{Lahman player id [chr]}
}}
\source{
Fantasy Pros Website
}
\usage{
data(pitchprojs_df)
}
\description{
Projected pitching stats (Fantasy Pros)
}
\keyword{datasets}
