% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{players_df}
\alias{players_df}
\title{Player Information}
\format{A \code{"players", 'tbl_df', 'tbl', "data.frame"} with 2444 rows and 8 columns
 \describe{
  \item{player_id}{Lahman id [chr]}
  \item{retro_id}{Retrosheet id [chr]}
  \item{bbref_id}{Baseball reference id [chr]}
  \item{first}{First name [chr]}
  \item{last}{Last name [chr]}
  \item{given}{Given name [chr]}
  \item{last_game}{Last game played in [date]}
  \item{full}{Full name [chr]}
}}
\source{
Lathman (see create_data_objects.R scripts)
}
\usage{
data(players_df)
}
\description{
Unique IDs, player names and last game information
}
\keyword{datasets}
