% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungeData.R
\name{prepMatchDup}
\alias{prepMatchDup}
\title{prepMatchDup}
\usage{
prepMatchDup(match_df, dup_df)
}
\arguments{
\item{match_df}{data.frame of duplicate matches}

\item{dup_df}{data.frame of team and positions of players}
}
\value{
data.frame containing player name, team, position
}
\description{
Fix duplicate matches
}
