% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungeData.R
\name{prepMatchNone}
\alias{prepMatchNone}
\title{prepMatchNone}
\usage{
prepMatchNone(no_row, player_df)
}
\arguments{
\item{no_row}{None matched players}

\item{player_df}{Data.frame of player names and info}
}
\value{
data.frame containing player name, team, position
}
\description{
Matches on last name only
}
