% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectStats.R
\name{projectErrors}
\alias{projectErrors}
\title{Project errors}
\usage{
projectErrors(fielding_df, batprojs_df, proj_year)
}
\arguments{
\item{fielding_df}{Fielding stats (data(fielding_df))}

\item{batprojs_df}{Batting projections (data(batprojs_df))}

\item{proj_year}{Year of projection}
}
\value{
Data.frame of player_id and projected errors
}
\description{
Project errors based on last season and position averages
}
