% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectStats.R
\name{projectGIDP}
\alias{projectGIDP}
\title{Project ground into double plays}
\usage{
projectGIDP(batting_df, batprojs_df, proj_year)
}
\arguments{
\item{batting_df}{Fielding stats (data(fielding_df))}

\item{batprojs_df}{Batting projections (data(batprojs_df))}

\item{proj_year}{Year of projection}
}
\value{
Data.frame of player_id and projected GIDP
}
\description{
Project GIDP based on last season and position averages
}
