% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectStats.R
\name{projectHolds}
\alias{projectHolds}
\title{Project holds}
\usage{
projectHolds(pitching_df, pitchprojs_df, proj_year)
}
\arguments{
\item{pitching_df}{Pitching stats (data(pitching_df))}

\item{pitchprojs_df}{Pitching projections (data(pitchprojs_df))}

\item{proj_year}{Year of projection}
}
\value{
Data.frame of player_id and projected Holds
}
\description{
Project holds based on last season and simple linear model
}
