% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{rankings_df}
\alias{rankings_df}
\title{Draft Rankings}
\format{A \code{"rankings", 'tbl_df', 'tbl', "data.frame"} data.frame with 2,529 rows 
and 10 variables
\describe{
  \item{player}{Standard player name [chr]}
  \item{team}{Players team (ESPN code) [chr]}
  \item{pos}{Player position(s) [chr]}
  \item{year}{Year of data [int]}
  \item{rank}{Overall player ranking [int]}
  \item{best}{Best draft position [int]}
  \item{worst}{Worst draft position [int]}
  \item{avg}{Average drafted position [dbl]}
  \item{std.dev}{St. Dev of average draft position [dbl]}
  \item{adp}{Average draft position (rounded and rank ordered) [dbl]}
  \item{player_id}{Lahman player id [chr]}
}}
\source{
Fantasy Pros Website
}
\usage{
data(rankings_df)
}
\description{
Projected draft rankings (Fantasy Pros)
}
\keyword{datasets}
