% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customRankings.R
\name{rrvToRankings}
\alias{rrvToRankings}
\title{Convert RRV data (and other relative measure) to a set of rankings}
\usage{
rrvToRankings(rrv_obj, projs_, configs, ...)
}
\arguments{
\item{rrv_obj}{RRV data from crRelativeRotisserieValue()}

\item{projs_}{Projection list of bat and pitch projections}

\item{configs}{an `fbbConfigs` object}
}
\value{
A 'draftRankings' object
}
\description{
Convert RRV data (and other relative measure) to a set of rankings
}
