% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectData.R
\name{scrapeBRStats}
\alias{scrapeBRStats}
\title{Scrape RetroSheet Annual Stats}
\usage{
scrapeBRStats(code, year)
}
\arguments{
\item{code}{Type of table to import (see data(baseballref_df))}

\item{year}{Year of data to grab}
}
\value{
Table of data for this type from this year
}
\description{
Get annual player statistic by player id
}
