% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseballreferenceScraper.R
\name{scrapeMarcelProjections}
\alias{scrapeMarcelProjections}
\title{Get Marcel projections}
\usage{
scrapeMarcelProjections(year = 2018)
}
\arguments{
\item{year}{year of stats}
}
\value{
Stats for that year
}
\description{
Scrape the Marcel 2019 projections from baseball reference
}
