% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCollection.R
\name{scrapeRetroAnnualData}
\alias{scrapeRetroAnnualData}
\title{Scrape RetroSheet Annual Stats}
\usage{
scrapeRetroAnnualData(playerid, years = 2018)
}
\arguments{
\item{playerid}{'Retrosheet' player id}

\item{years}{Years of data to grab}
}
\value{
List of pitching, batting and fielding annual stats
}
\description{
Get annual player statistic by player id
}
