% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonAnalysis.R
\name{seasonAnalysis}
\alias{seasonAnalysis}
\title{Analyze Season to get point differntials}
\usage{
seasonAnalysis(season, ...)
}
\arguments{
\item{season}{A `season` object}
}
\value{
Median difference between categories
}
\description{
Get point differentials across categories
}
