% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{seasons_}
\alias{seasons_}
\title{Season Results}
\format{A list with individual \code{"seasons", 'tbl_df', 'tbl', "data.frame"} 
data.frames with varying rows and variables.  Most common variables are:
 \describe{
  \item{manager}{Unique team manager (ESPN) id [int]}
  \item{rank}{Season Ranking [int]}
  \item{r}{Runs accumulated [int]}
  \item{hr}{Home runs accumulated [int]}
  \item{rbi}{RBIs accumulated [int]}
  \item{so}{Strikeouts (batter) accumulated [int]}
  \item{sb}{Stolen bases accumulated [int]}
  \item{obp}{On base percentage accumulated [num]}
  \item{slg}{Slugging percentage accumulated [num]}
  \item{gidp}{Grounded into double plays accumulated [int]}
  \item{e}{Errors accumulated [int]}
  \item{ip}{Innings pitched accumulated [int]}
  \item{qs}{Quality starts accumulated [int]}
  \item{k}{Strikeouts (pitcher) accumulated [int]}
  \item{sv}{Saves accumulated [int]}
  \item{hd}{Holds accumulated [int]}
  \item{whip}{WHIP accumulated [num]}
  \item{moves}{Moves made during the season [int]}
  \item{starts}{Pitchers starts used [int]}
}}
\source{
Author's own Leagues
}
\usage{
data(seasons_)
}
\description{
Previous Season Results from 2015 to 2018
}
\keyword{datasets}
