% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customRankings.R
\name{simBattingSeason}
\alias{simBattingSeason}
\title{Simulate a batting season}
\usage{
simBattingSeason(iter, rosterstats_df, configs)
}
\arguments{
\item{iter}{Iteration number of the simulation}

\item{rosterstats_df}{Individual player stats or projections}

\item{configs}{An `fbbConfigs` object}
}
\value{
Summed total of hitting statistics/projections
}
\description{
Sum up all batting stats or projections to get total rotisserie raw category counts and
ratios
}
