% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSeasons.R
\name{simulateFBB}
\alias{simulateFBB}
\title{Calculate SLG}
\usage{
simulateFBB(configs, verbose = 1, ...)
}
\arguments{
\item{x_df}{Data.frame}
}
\value{
Vector of slugging percentages
}
\description{
Calculate slugging percentage
}
