% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R
\docType{data}
\name{teams_df}
\alias{teams_df}
\title{Team Identifier Codes}
\format{A \code{"seasons", 'tbl_df', 'tbl', "data.frame"} with 30 rows and 6 columns
 \describe{
  \item{fp_id}{Fantasy Pros team id [chr]}
  \item{br_id}{Baseball reference team id [chr]}
  \item{lahman_id}{Lahman team id [chr]}
  \item{retro_id}{Retro sheet standard Pros team id [chr]}
  \item{espn_id}{ESPN team id [chr]}
  \item{retrosheet_id}{Retro sheet data scraping team id [chr]}
}}
\source{
Author compiled
}
\usage{
data(teams_df)
}
\description{
Team codes from various sources
}
\keyword{datasets}
