% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customRankings.R
\name{customRankings.rrv}
\alias{customRankings.rrv}
\title{Relative rotisserie values (rrv)}
\usage{
\method{customRankings}{rrv}(type, configs, ...)
}
\arguments{
\item{type}{Type of custom ranking system to use ('xxx' for now)}

\item{configs}{An `fbbConfigs` object}

\item{verbose}{[1] The level of reponse statements you'd like (higher = more)}
}
\value{
A custom ranking set
}
\description{
Custom rankings that takes into account relative rotisserie point values
}
